refactor(cli): split list command by tab/recents views
